APG-1222: Add the nDelius API Integration client methods for getting the manager associated with a Requirement or Licence Condition
